Refactor the codebase to clean it up, use logging, move comments to docstrings, etc.
I'm a terrible person. This commit is the nuclear option, not atomic.
Sorry about that. Basically, this is the resolution of a lot of WIP code
that I've been hacking on for a few weeks, as I built up my
understanding of the code base and how the communication with the
Arduino was supposed to work. Since it wasn't working , initially, I
didn't feel comfortable commiting, until it did. Now it works. In doing
so, I changed the message interface between the Python code (the RPi) and the
Arduino.